Retry when API request failed
