+++ commands.go
@@ -1,10 +1,8 @@
 package main
 
 import (
-	"bufio"
 	"fmt"
 	"os"
-	"strconv"
 	"strings"
 	"text/template"
 	"time"
@@ -132,22 +130,6 @@ var commandUpdate = cli.Command{
 			Usage: "Update rolefullname.",
 		},
 		cli.BoolFlag{Name: "overwriteRoles, o", Usage: "Overwrite roles instead of adding specified roles."},
-	},
-}
-
-var commandThrow = cli.Command{
-	Name:      "throw",
-	Usage:     "Post metric values",
-	ArgsUsage: "[--host | -H <hostId>] [--service | -s <service>] stdin",
-	Description: `
-    Post metric values to 'host metric' or 'service metric'.
-    Output format of metric values are compatible with that of a Sensu plugin.
-    Requests "POST /api/v0/tsdb". See https://mackerel.io/api-docs/entry/host-metrics#post .
-`,
-	Action: doThrow,
-	Flags: []cli.Flag{
-		cli.StringFlag{Name: "host, H", Value: "", Usage: "Post host metric values to <hostID>."},
-		cli.StringFlag{Name: "service, s", Value: "", Usage: "Post service metric values to <service>."},
 	},
 }
 
@@ -407,71 +389,6 @@ func doUpdate(c *cli.Context) error {
 		}
 
 		logger.Log("updated", hostID)
-	}
-	return nil
-}
-
-func doThrow(c *cli.Context) error {
-	optHostID := c.String("host")
-	optService := c.String("service")
-
-	var metricValues []*(mkr.MetricValue)
-
-	scanner := bufio.NewScanner(os.Stdin)
-	for scanner.Scan() {
-		line := scanner.Text()
-
-		// name, value, timestamp
-		// ex.) tcp.CLOSING 0 1397031808
-		items := strings.Fields(line)
-		if len(items) != 3 {
-			continue
-		}
-		value, err := strconv.ParseFloat(items[1], 64)
-		if err != nil {
-			logger.Log("warning", fmt.Sprintf("Failed to parse values: %s", err))
-			continue
-		}
-		time, err := strconv.ParseInt(items[2], 10, 64)
-		if err != nil {
-			logger.Log("warning", fmt.Sprintf("Failed to parse values: %s", err))
-			continue
-		}
-
-		name := items[0]
-		if optHostID != "" && !strings.HasPrefix(name, "custom.") {
-			name = "custom." + name
-		}
-
-		metricValue := &mkr.MetricValue{
-			Name:  name,
-			Value: value,
-			Time:  time,
-		}
-
-		metricValues = append(metricValues, metricValue)
-	}
-	logger.ErrorIf(scanner.Err())
-
-	client := newMackerelFromContext(c)
-
-	if optHostID != "" {
-		err := client.PostHostMetricValuesByHostID(optHostID, metricValues)
-		logger.DieIf(err)
-
-		for _, metric := range metricValues {
-			logger.Log("thrown", fmt.Sprintf("%s '%s\t%f\t%d'", optHostID, metric.Name, metric.Value, metric.Time))
-		}
-	} else if optService != "" {
-		err := client.PostServiceMetricValues(optService, metricValues)
-		logger.DieIf(err)
-
-		for _, metric := range metricValues {
-			logger.Log("thrown", fmt.Sprintf("%s '%s\t%f\t%d'", optService, metric.Name, metric.Value, metric.Time))
-		}
-	} else {
-		cli.ShowCommandHelp(c, "throw")
-		os.Exit(1)
 	}
 	return nil
 }
